Remove create_pool_shared
It is no longer shared.
