+++ src/dbus_api/api/manager_3_0/methods.rs
@@ -2,20 +2,28 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-use std::convert::TryFrom;
+use std::{convert::TryFrom, path::Path};
 
-use dbus::Message;
+use dbus::{arg::Array, Message};
 use dbus_tree::{MTSync, MethodInfo, MethodResult};
 
 use crate::{
     dbus_api::{
-        api::shared::{create_pool_shared, set_key_shared, unlock_pool_shared},
+        api::shared::{set_key_shared, unlock_pool_shared},
+        blockdev::create_dbus_blockdev,
         consts,
+        pool::create_dbus_pool,
         types::{DbusErrorEnum, TData, OK_STRING},
-        util::{engine_to_dbus_err_tuple, get_next_arg},
+        util::{engine_to_dbus_err_tuple, get_next_arg, tuple_to_option},
     },
-    engine::{DeleteAction, KeyDescription, MappingDeleteAction, PoolUuid},
+    engine::{
+        CreateAction, DeleteAction, EncryptionInfo, KeyDescription, MappingDeleteAction, Name,
+        PoolUuid,
+    },
+    stratis::StratisError,
 };
+
+type EncryptionParams = (Option<(bool, String)>, Option<(bool, (String, String))>);
 
 pub fn destroy_pool(m: &MethodInfo<MTSync<TData>, TData>) -> MethodResult {
     let message: &Message = m.msg;
@@ -156,5 +164,94 @@ pub fn engine_state_report(m: &MethodInfo<MTSync<TData>, TData>) -> MethodResult
 }
 
 pub fn create_pool(m: &MethodInfo<MTSync<TData>, TData>) -> MethodResult {
-    create_pool_shared(m)
+    let message: &Message = m.msg;
+    let mut iter = message.iter_init();
+
+    let name: &str = get_next_arg(&mut iter, 0)?;
+    let redundancy_tuple: (bool, u16) = get_next_arg(&mut iter, 1)?;
+    let devs: Array<&str, _> = get_next_arg(&mut iter, 2)?;
+    let (key_desc_tuple, clevis_tuple): EncryptionParams = (
+        Some(get_next_arg(&mut iter, 3)?),
+        Some(get_next_arg(&mut iter, 4)?),
+    );
+
+    let return_message = message.method_return();
+
+    let default_return: (bool, (dbus::Path<'static>, Vec<dbus::Path<'static>>)) =
+        (false, (dbus::Path::default(), Vec::new()));
+
+    let key_desc = match key_desc_tuple.and_then(tuple_to_option) {
+        Some(kds) => match KeyDescription::try_from(kds) {
+            Ok(kd) => Some(kd),
+            Err(e) => {
+                let (rc, rs) = engine_to_dbus_err_tuple(&e);
+                return Ok(vec![return_message.append3(default_return, rc, rs)]);
+            }
+        },
+        None => None,
+    };
+
+    let clevis_info = match clevis_tuple.and_then(tuple_to_option) {
+        Some((pin, json_string)) => match serde_json::from_str(json_string.as_str()) {
+            Ok(j) => Some((pin, j)),
+            Err(e) => {
+                let (rc, rs) = engine_to_dbus_err_tuple(&StratisError::Serde(e));
+                return Ok(vec![return_message.append3(default_return, rc, rs)]);
+            }
+        },
+        None => None,
+    };
+
+    let object_path = m.path.get_name();
+    let dbus_context = m.tree.get_data();
+    let mut mutex_lock = dbus_context.engine.blocking_lock();
+    let result = log_action!(mutex_lock.create_pool(
+        name,
+        &devs.map(|x| Path::new(x)).collect::<Vec<&Path>>(),
+        tuple_to_option(redundancy_tuple),
+        &EncryptionInfo {
+            key_description: key_desc,
+            clevis_info,
+        }
+    ));
+
+    let msg = match result {
+        Ok(pool_uuid_action) => {
+            let results = match pool_uuid_action {
+                CreateAction::Created(uuid) => {
+                    let (_, pool) = get_pool!(mutex_lock; uuid; default_return; return_message);
+
+                    let pool_object_path: dbus::Path = create_dbus_pool(
+                        dbus_context,
+                        object_path.clone(),
+                        &Name::new(name.to_string()),
+                        uuid,
+                        pool,
+                    );
+
+                    let bd_paths = pool
+                        .blockdevs()
+                        .into_iter()
+                        .map(|(uuid, tier, bd)| {
+                            create_dbus_blockdev(
+                                dbus_context,
+                                pool_object_path.clone(),
+                                uuid,
+                                tier,
+                                bd,
+                            )
+                        })
+                        .collect::<Vec<_>>();
+                    (true, (pool_object_path, bd_paths))
+                }
+                CreateAction::Identity => default_return,
+            };
+            return_message.append3(results, DbusErrorEnum::OK as u16, OK_STRING.to_string())
+        }
+        Err(x) => {
+            let (rc, rs) = engine_to_dbus_err_tuple(&x);
+            return_message.append3(default_return, rc, rs)
+        }
+    };
+    Ok(vec![msg])
 }
